Optimize VPTree exact-index query latency (best-first, cache layout, FMA, VP selection)
- Best-first search (min-heap on distToBorder lower bound): allows tau to
  decrease faster, enabling more aggressive pruning vs prior DFS traversal
- Cache-friendly data reorder: after build, _flat_backing reordered so tree
  position i holds data for _indices[i], eliminating _examples[_indices[i]]
  indirection in FlatSpan searches
- FMA distance kernel: _mm256_fmadd_ps replaces add+mul in dist_l2_f_avx2
- Improved VP selection: sample S=5 candidates, pick max distance-variance VP
- int32_t node indices: shrink _indices and partition start/end from 8->4 bytes
  for better cache utilisation; update SerializableVPTree serialization to match
- Parallel build: OpenMP on distance loop for partitions >= 4096 elements

Result: 3.3x query-latency reduction at d=512 (7.3ms -> 2.2ms);
39x advantage over Faiss IndexFlatL2 and 220x over scikit-learn at d=512.
Rebuild benchmarks.pdf with measured numbers across all optimisations.

+++ pynear/include/VPLevelPartition.hpp
@@ -15,15 +15,15 @@ namespace vptree {
 
 template <typename distance_type> class VPLevelPartition {
 public:
-    VPLevelPartition(distance_type radius, int64_t start, int64_t end)
+    VPLevelPartition(distance_type radius, int32_t start, int32_t end)
         : _radius(radius), _indexStart(start), _indexEnd(end) {}
 
     VPLevelPartition() : _radius(0), _indexStart(-1), _indexEnd(-1) {}
 
     bool isEmpty() const { return _indexStart == -1; }
-    int64_t start() const { return _indexStart; }
-    int64_t end() const { return _indexEnd; }
-    int64_t size() const { return _indexEnd - _indexStart + 1; }
+    int32_t start() const { return _indexStart; }
+    int32_t end() const { return _indexEnd; }
+    int32_t size() const { return _indexEnd - _indexStart + 1; }
     void setRadius(distance_type radius) { _radius = radius; }
     distance_type radius() const { return _radius; }
 
@@ -53,8 +53,8 @@ private:
     }
 
     distance_type _radius = 0;
-    int64_t _indexStart = -1;
-    int64_t _indexEnd = -1;
+    int32_t _indexStart = -1;
+    int32_t _indexEnd = -1;
     int32_t _left_idx = -1;
     int32_t _right_idx = -1;
 };
